+++ CHANGELOG.md
@@ -6,9 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.1] - 2023-11-03
+
 ### Changed
 
 - Peripheral proxy now directional
+
+### Fixed
+
+- Integrated dynamic integration [#28]
 
 ## [1.4.0] - 2023-09-14